backend/requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Utilities
jinja2>=3.1.2
python-dateutil>=2.8.2
uuid6>=2024.1.12

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0

# Optional: HTML parsing for scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
